Went back to unique callsite IDs (not sure why they were dropped in the first place).

+++ instrument/src/CallInstHandler.h
@@ -22,7 +22,7 @@ class CallInstHandler : public TimestampPlacerHandler
 	void addIgnore(std::string func_name);
 
     private:
-    uint32_t call_idx;
+    uint64_t call_idx;
     PassLog& log;
     llvm::Function* link_arg_const_func;
     llvm::Function* link_arg_func;
